pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "gsheeter"
authors = [{ name = "Yunjong Guk", email = "haydenkuk@gmail.com" }]
maintainers = []
readme = "README.md"
keywords = ["spreadsheets", "google-spreadsheets", "google-sheets"]
classifiers = [
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Intended Audience :: End Users/Desktop",
  "Intended Audience :: Science/Research",
  "Topic :: Office/Business :: Financial :: Spreadsheet",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "asttokens==3.0.0",
  "build==1.2.2.post1",
  "cachetools==5.5.2",
  "certifi==2025.1.31",
  "charset-normalizer==3.4.1",
  "colorama==0.4.6",
  "dateparser==1.2.1",
  "et_xmlfile==2.0.0",
  "executing==2.2.0",
  "google-api-core==2.24.2",
  "google-api-python-client==2.166.0",
  "google-auth==2.38.0",
  "google-auth-httplib2==0.2.0",
  "google-auth-oauthlib==1.2.1",
  "googleapis-common-protos==1.69.2",
  "httplib2==0.22.0",
  "idna==3.10",
  "numpy==2.2.4",
  "oauthlib==3.2.2",
  "openpyxl==3.1.5",
  "packaging==25.0",
  "pandas==2.2.3",
  "proto-plus==1.26.1",
  "protobuf==6.30.2",
  "pyasn1==0.6.1",
  "pyasn1_modules==0.4.2",
  "Pygments==2.19.1",
  "pyparsing==3.2.3",
  "pyproject_hooks==1.2.0",
  "python-dateutil==2.9.0.post0",
  "pytz==2025.2",
  "regex==2024.11.6",
  "requests==2.32.3",
  "requests-oauthlib==2.0.0",
  "rsa==4.9",
  "six==1.17.0",
  "typing_extensions==4.13.2",
  "tzdata==2025.2",
  "tzlocal==5.3.1",
  "uritemplate==4.1.1",
  "urllib3==2.4.0",
  "XlsxWriter==3.2.3"
]

requires-python = ">=3.9"
dynamic = ["version", "description"]

[project.urls]
Documentation = ""
Source = "https://github.com/haydenkuk/gsheeter"
Issues = "https://github.com/haydenkuk/gsheeter/issues"

[project.license]
file = "LICENSE.txt"

[tool.flit.sdist]
include = ["docs/", "HISTORY.rst"]

[tool.setuptools.package-data]
gsheeter = ["gsheeter/py.typed"]